CascadeProjects/imageboard/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
